+++ Sources/QuillCodeCore/ChatModels.swift
@@ -1,0 +1,37 @@
+import Foundation
+
+public enum ChatRole: String, Codable, Sendable {
+    case system
+    case user
+    case assistant
+    case tool
+}
+
+public struct ChatMessage: Codable, Sendable, Hashable, Identifiable {
+    public let id: UUID
+    public var role: ChatRole
+    public var content: String
+    public var createdAt: Date
+
+    public init(id: UUID = UUID(), role: ChatRole, content: String, createdAt: Date = Date()) {
+        self.id = id
+        self.role = role
+        self.content = content
+        self.createdAt = createdAt
+    }
+}
+
+public enum MessageFeedbackValue: String, Codable, Sendable, Hashable {
+    case helpful
+    case notHelpful
+}
+
+public struct MessageFeedback: Codable, Sendable, Hashable {
+    public var messageID: UUID
+    public var value: MessageFeedbackValue
+
+    public init(messageID: UUID, value: MessageFeedbackValue) {
+        self.messageID = messageID
+        self.value = value
+    }
+}
