+++ Sources/QuillCodeCore/JSONHelpers.swift
@@ -1,0 +1,16 @@
+import Foundation
+
+public enum JSONHelpers {
+    public static func encodePretty<T: Encodable>(_ value: T) throws -> String {
+        let encoder = JSONEncoder()
+        encoder.outputFormatting = [.prettyPrinted, .sortedKeys]
+        encoder.dateEncodingStrategy = .iso8601
+        return String(decoding: try encoder.encode(value), as: UTF8.self)
+    }
+
+    public static func decode<T: Decodable>(_ type: T.Type, from string: String) throws -> T {
+        let decoder = JSONDecoder()
+        decoder.dateDecodingStrategy = .iso8601
+        return try decoder.decode(type, from: Data(string.utf8))
+    }
+}
