+++ Sources/QuillCodeCore/ThreadEventModels.swift
@@ -1,0 +1,36 @@
+import Foundation
+
+public enum ThreadEventKind: String, Codable, Sendable {
+    case message
+    case toolQueued
+    case toolRunning
+    case toolCompleted
+    case toolFailed
+    case approvalRequested
+    case approvalDecided
+    case reviewComment
+    case notice
+    case messageFeedback
+}
+
+public struct ThreadEvent: Codable, Sendable, Hashable, Identifiable {
+    public var id: UUID
+    public var kind: ThreadEventKind
+    public var createdAt: Date
+    public var summary: String
+    public var payloadJSON: String?
+
+    public init(
+        id: UUID = UUID(),
+        kind: ThreadEventKind,
+        createdAt: Date = Date(),
+        summary: String,
+        payloadJSON: String? = nil
+    ) {
+        self.id = id
+        self.kind = kind
+        self.createdAt = createdAt
+        self.summary = summary
+        self.payloadJSON = payloadJSON
+    }
+}
