+++ docs/DECISIONS.md
@@ -101,8 +101,9 @@
 - TrustedRouter action and safety transports are separate files. `TrustedRouterLLMClient` owns action streaming; `TrustedRouterSafetyModelClient` owns Auto-review model calls. They share API-key resolution through `TrustedRouterAPIKeyResolver` and JSON response payloads through `TrustedRouterChatParameters` so neither transport depends on the other.
 - Static Auto safety fallback policy lives in `StaticSafetyPolicy`, not `StaticSafetyReviewer`. The reviewer owns mode behavior, hard-deny precedence, low-risk approval, and model-backed fallback orchestration; the policy owns table-driven hard-deny patterns, generic user-intent rules, and pull-request-specific intent routing.
 - Static Auto safety intent rules must be tool-limited. Broad user words like `run`, `execute`, `disk`, `storage`, `whoami`, and `openclaw` should approve the bounded tool families that satisfy that request, not any append/destructive tool the model happens to propose.
-- TrustedRouter model defaults and catalog normalization live in focused core files, not the general `Models.swift` domain bucket. `TrustedRouterDefaults.swift` owns provider IDs, aliases, bundled fallback rows, Nike 1.0/Synth branding, and sort/category policy; `ModelInfo.swift` owns the catalog value records.
-- App configuration is a focused core boundary. `AppConfig.swift` owns persisted model/mode/base-URL settings, OAuth/developer-override compatibility, signed-in TrustedRouter account metadata, and favorite-model normalization; `Models.swift` stays focused on general thread/project records.
+- Core domain models live in focused files rather than a general `Models.swift` bucket. `AgentMode.swift`, `ChatModels.swift`, `AgentPlanModels.swift`, `SubagentModels.swift`, `ApprovalModels.swift`, `ThreadEventModels.swift`, `MemoryModels.swift`, `ChatThread.swift`, and `JSONHelpers.swift` own the general thread/chat/memory domain; parity gates prevent the old umbrella file from returning.
+- TrustedRouter model defaults and catalog normalization live in focused core files. `TrustedRouterDefaults.swift` owns provider IDs, aliases, bundled fallback rows, Nike 1.0/Synth branding, and sort/category policy; `ModelInfo.swift` owns the catalog value records.
+- App configuration is a focused core boundary. `AppConfig.swift` owns persisted model/mode/base-URL settings, OAuth/developer-override compatibility, signed-in TrustedRouter account metadata, and favorite-model normalization; general chat/thread/project models should not own settings compatibility rules.
 - Core tool records are a focused schema boundary. `ToolModels.swift` owns `ToolDefinition`, `ToolCall`, `ToolResult`, built-in core tool definitions, tool-call transcript redaction, and tool-specific browser/memory output compatibility; general chat/thread/project models should not own tool payload compatibility rules.
 - Core automation records are a focused scheduling boundary. `AutomationModels.swift` owns automation kind/status/schedule enums, recurrence interval semantics, next-run calculation, and display sorting; general chat/thread/project models should not own recurring-workflow scheduling rules.
 - Core project records are a focused workspace boundary. `ProjectModels.swift` owns local/SSH connection parsing and display, project refs, instructions, local environment actions, and project extension manifests; general thread/message models should not own workspace connection or extension compatibility.
